requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==19.3.0
Automat==0.8.0
autopep8==1.6.0
blinker==1.4
blobfile==2.0.1
build==0.10.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.3
cloud-init==21.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.7
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.9.0
Flask==2.0.2
fonttools==4.38.0
frozenlist==1.3.3
httplib2==0.14.0
hyperlink==19.0.0
idna==3.3
importlib-metadata==1.5.0
incremental==16.10.1
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.2.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==4.9.2
macaroonbakery==1.3.1
MarkupSafe==2.0.1
matplotlib==3.6.3
more-itertools==4.2.0
multidict==6.0.4
netifaces==0.10.4
numpy==1.21.3
oauthlib==3.1.0
openai==0.19.0
openpyxl==3.0.9
packaging==23.0
pandas==1.3.4
pandas-stubs==1.2.0.35
pexpect==4.6.0
Pillow==9.4.0
pip-tools==6.12.2
plotly==5.13.0
protobuf==3.6.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycodestyle==2.8.0
pycryptodomex==3.17
pycups==1.9.73
pydantic==1.10.4
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.3.1
regex==2022.10.31
replicate==0.4.0
requests==2.26.0
requests-unixsocket==0.2.0
scikit-learn==1.2.1
scipy==1.10.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
sklearn==0.0.post1
sos==4.1
spotipy==2.19.0
ssh-import-id==5.10
systemd-python==234
tenacity==8.2.1
threadpoolctl==3.1.0
tiktoken==0.2.0
toml==0.10.2
tomli==2.0.1
tqdm==4.62.3
Twisted==18.9.0
typing_extensions==4.4.0
ubuntu-advantage-tools==27.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.7
wadllib==1.3.3
watchdog==2.2.1
Werkzeug==2.0.2
yarl==1.8.2
zipp==1.0.0
zope.interface==4.7.1
